File helps calculating CNN receptive field
This is a really simple way to help you understand how CNNs work, in terms of receptive field and stride. 
This is a simple file, it does not take into account things like different padding and skip connections. 
It also assumes "full" padding as Keras uses, as opposed to "valid" padding. 
Values of 1 are hidden by sheet styling but are required for the proper calculations. 

For normal Pooling layers, say a 2x2 pool, this is a kernel size 2 with a stride/pool of 2. 
For a normal Conv2D layer, say 3x3, this is kernel size 3 with stride/pool of 1.